+++ spec/doc_parser_spec.rb
@@ -1,0 +1,7 @@
+require 'spec_helper'
+
+describe LinkThumbnailer::DocParser do
+
+  it { should respond_to :parse }
+
+end