+++ spec/opengraph_spec.rb
@@ -1,0 +1,7 @@
+require 'spec_helper'
+
+describe LinkThumbnailer::Opengraph do
+
+  it { LinkThumbnailer::Opengraph.should respond_to :parse }
+
+end