+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 0.0.4
+
+- Add specs for almost all classes
+- Add a method `to_json` for WebImage class to be able to get a usable array of images' attributes
+
+# 0.0.3
+
+- Add specs for LinkThumbnailer class
+- Refactor config system, now using dedicated configuration class
+
 # 0.0.2
 
 - Added Rspec